requirements.txt
docker==3.5.0
PyYAML==3.12
